Move utility functions into separate files
This avoid circular dependency issues.
Over time, we'll move more functions into
their own files like this.
