+++ _tools/run/helpers/merge/index.js
@@ -8,7 +8,7 @@ const fsPath = require('path')
 const fsPromises = require('fs/promises')
 
 // Local helpers
-const htmlFilePaths = require('../helpers').htmlFilePaths
+const htmlFilePaths = require('../paths/htmlFilePaths.js')
 
 // Make IDs in HTML unique by prefixing them
 // with the slug of the filename, and updating
